backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
python-jose
razorpay
reportlab
python-multipart
twilio